requirements.txt
streamlitrequestspandasplotlytransformerstorchnewsapi-pythonpython-dotenvyfinancepytz